requirements.txt
Flask==2.2.3
Flask-Bootstrap==3.3.7.1
Jinja2==3.1.2
MarkupSafe==2.1.2
PyYAML==5.4.1
visitor==0.1.3
Werkzeug==2.2.3
Flask-MySQLdb==1.0.1
passlib==1.7.4